tests/wpt/tests/keyboard-map/navigator-keyboard-map-two-parallel-requests.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

promise_test((t) => {
  const p1 = navigator.keyboard.getLayoutMap();
  const p2 = navigator.keyboard.getLayoutMap();
  // p1 and p2 should be the same promise instance.
  assert_equals(p1, p2);
  return Promise.all([p1, p2]);
}, '[Keyboard Map] getLayoutMap() twice in parallel');

</script>
